+++ content/abilities/validators/validator.gd
@@ -1,0 +1,9 @@
+
+class_name Validator
+
+func validate(origin: Vector2i, area: Array[Vector2i]):
+	var output: Array[Vector2i] = []
+	for cell in area:
+		if true:
+			output.append(cell)
+	return output