+++ gameplay2/battle/player_input_controller.gd
@@ -1,7 +1,5 @@
 extends Node
 
-class_name PlayerInputController
-
 @export var state_machine: StateMachine
 
 var idle_state